Reorganized the kernel/platform directory to better distinguish architectures from sub-architectures Added makefile rules for the emulation of RPi2 and RPi3

+++ kernel/platform/arm64/arm-cortex-a72/platform-defs.h
@@ -15,7 +15,7 @@ static inline void MemFence(){
 }
 
 struct spinlock{
-  uint32_t m;
+  uint64_t m;
 };
